+++ android/app/src/main/AndroidManifest.xml
@@ -26,12 +26,12 @@
             android:windowSoftInputMode="adjustResize"
             tools:ignore="DiscouragedApi,LockedOrientationActivity" />
         <activity
-            android:name=".presentation.joininfo.JoinInfoActivity"
+            android:name=".presentation.join.JoinActivity"
             android:exported="false"
             android:screenOrientation="portrait"
             tools:ignore="DiscouragedApi,LockedOrientationActivity" />
         <activity
-            android:name=".presentation.completion.JoinCompleteActivity"
+            android:name=".presentation.join.complete.JoinCompleteActivity"
             android:exported="false"
             android:screenOrientation="portrait"
             tools:ignore="DiscouragedApi,LockedOrientationActivity" />
