+++ android/app/src/main/java/com/woowacourse/ody/presentation/invitecode/InviteCodeActivity.kt
@@ -10,7 +10,7 @@ import com.google.android.material.snackbar.Snackbar
 import com.woowacourse.ody.R
 import com.woowacourse.ody.data.remote.ody.repository.DefaultMeetingRepository
 import com.woowacourse.ody.databinding.ActivityInviteCodeBinding
-import com.woowacourse.ody.presentation.joininfo.JoinInfoActivity
+import com.woowacourse.ody.presentation.join.JoinActivity
 import com.woowacourse.ody.presentation.common.listener.BackListener
 import com.woowacourse.ody.presentation.common.observeEvent
 
@@ -55,7 +55,7 @@ class InviteCodeActivity : AppCompatActivity(), BackListener {
 
     private fun navigateToJoinView() {
         val inviteCode = viewModel.inviteCode.value ?: return
-        startActivity(JoinInfoActivity.getIntent(inviteCode, this))
+        startActivity(JoinActivity.getIntent(inviteCode, this))
     }
 
     override fun onBack() = finish()
