+++ android/app/src/main/java/com/woowacourse/ody/presentation/meeting/MeetingInfoActivity.kt
@@ -14,14 +14,14 @@ import androidx.fragment.app.Fragment
 import com.woowacourse.ody.databinding.ActivityMeetingInfoBinding
 import com.woowacourse.ody.presentation.common.listener.BackListener
 import com.woowacourse.ody.presentation.common.ViewPagerAdapter
-import com.woowacourse.ody.presentation.completion.JoinCompleteActivity
+import com.woowacourse.ody.presentation.join.complete.JoinCompleteActivity
 import com.woowacourse.ody.presentation.meeting.complete.MeetingCompletionActivity
 import com.woowacourse.ody.presentation.meeting.date.MeetingDateFragment
 import com.woowacourse.ody.presentation.meeting.destination.MeetingDestinationFragment
 import com.woowacourse.ody.presentation.intro.IntroActivity
 import com.woowacourse.ody.presentation.meeting.name.MeetingNameFragment
-import com.woowacourse.ody.presentation.nickname.JoinNickNameFragment
-import com.woowacourse.ody.presentation.departure.DepartureFragment
+import com.woowacourse.ody.presentation.join.nickname.JoinNickNameFragment
+import com.woowacourse.ody.presentation.join.departure.DepartureFragment
 import com.woowacourse.ody.presentation.meeting.time.MeetingTimeFragment
 import com.woowacourse.ody.presentation.common.observeEvent
