+++ android/app/src/main/res/layout/activity_join.xml
@@ -21,7 +21,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.joininfo.JoinInfoActivity">
+        tools:context=".presentation.join.JoinActivity">
 
         <include
             android:id="@+id/tb_join_info"
